Moving common query functions to a base class
