+++ lib/openstack_api/openstack_user.py
@@ -6,10 +6,10 @@ from openstack_api.dataclasses import QueryParams
 from openstack_api.openstack_connection import OpenstackConnection
 from openstack_api.openstack_identity import OpenstackIdentity
 from openstack_api.openstack_query import OpenstackQuery
-from openstack_api.openstack_wrapper_base import OpenstackWrapperBase
+from openstack_api.openstack_query_base import OpenstackQueryBase
 
 
-class OpenstackUser(OpenstackWrapperBase):
+class OpenstackUser(OpenstackQueryBase):
     # Lists all possible query presets for user.list
     SEARCH_QUERY_PRESETS: List[str] = [
         "all_users",
@@ -26,29 +26,12 @@ class OpenstackUser(OpenstackWrapperBase):
         self._identity_api = OpenstackIdentity(self._connection_cls)
         self._query_api = OpenstackQuery(self._connection_cls)
 
-    def __getitem__(self, item):
-        return getattr(self, item)
-
     def _get_query_property_funcs(self, _) -> Dict[str, Callable[[Any], Any]]:
         """
         Returns property functions for use with OpenstackQuery.parse_properties
         :param cloud_account: The associated clouds.yaml account
         """
         return {}
-
-    def search(self, cloud_account: str, query_params: QueryParams, **kwargs):
-        """
-        Performs a search of users and returns table of results
-        :param cloud_account: The associated clouds.yaml account
-        :param query_params: See QueryParams
-        """
-        return self._query_api.search_resource(
-            cloud_account=cloud_account,
-            search_api=self,
-            property_funcs=self._get_query_property_funcs(cloud_account),
-            query_params=query_params,
-            **kwargs,
-        )
 
     def search_all_users(self, cloud_account: str, user_domain: str, **_) -> List[User]:
         """
